data_flow_and_processing/stream_processing/stream-stock/requirements.txt
kafka-python
faust
sqlalchemy
streamlit
fastapi